split files into http, http1, and http2
